% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_boot.R
\name{fitness_boot}
\alias{fitness_boot}
\title{fitness_boot}
\usage{
fitness_boot(
  model,
  nboot,
  index,
  vectors = NULL,
  dim = NULL,
  data,
  amat,
  newdata,
  modmat.new = NULL,
  renewdata = NULL,
  criterion = c("AIC", "BIC", "LRT"),
  alpha = 0.05,
  fit.name = NULL,
  method = c("eigen", "1d"),
  quiet = FALSE,
  corenum = NULL
)
}
\arguments{
\item{model}{An aster model object.}

\item{nboot}{The number of bootstrap iterations desired.}

\item{index}{The indices denoting which components of the canonical parameter vector are parameters of interest.}

\item{vectors}{The indices denoting which reducing subspace of Fisher information is desired to construct envelope estimators. \cr
Must be specified if \code{method = "eigen"}.}

\item{dim}{The dimension of the envelope space used to construct envelope estimators.\cr
Must be specified if \code{method = "1d"}.}

\item{data}{An asterdata object corresponding to the original data.}

\item{amat}{This object can either be an array or a matrix.\cr
It specifies a linear combination of mean-value parameters that correspond to expected Darwinian fitness.\cr
See the \code{aster} function help page in the original \code{aster} package for more details.}

\item{newdata}{A dataframe corresponding to hypothetical individuals in which expected Darwinian fitness is to be estimated.}

\item{modmat.new}{A model matrix corresponding to hypothetical individuals in which expected Darwinian fitness is to be estimated.}

\item{renewdata}{A dataframe in long format corresponding to hypothetical individuals in which expected Darwinian fitness is to be estimated.}

\item{criterion}{A model selection criterion of choice.}

\item{alpha}{The type 1 error rate desired for the LRT.}

\item{fit.name}{An expression that appears in the name of the nodes that correspond to Darwinian fitness.\cr
This is only necessary if \code{renewdata} is not provided.}

\item{method}{The procedure used to obtain envelope estimators.}

\item{quiet}{A logical argument. If FALSE, the function displays how much time it takes to run \code{m} iterations.}

\item{corenum}{The number of cores specified for speeding up the bootstrap process.}
}
\value{
a list containing the following elements in order:
\item{env.boot.out}{Estimated expected Darwinian fitness from generated data obtained from Steps 3a-3d in the bootstrap procedure using the envelope
estimator constructed using reducing subspaces. }
\item{MLE.boot.out}{Estimated expected Darwinian fitness from generated data obtained from Steps 3a-3d in the bootstrap procedure using the MLE. }
\item{MLE.tau.boot}{Estimated mean-value parameter vectors from generated data obtained from Steps 3a-3d in the bootstrap procedure using the MLE. }
\item{env.tau.boot}{Estimated mean-value parameter vectors from generated data obtained from Steps 3a-3d in the bootstrap procedure using the envelope estimator constructed using the 1d algorithm. }
\item{P.list}{A list of all estimated projections into the envelope space constructed from reducing subspaces for Steps 3a-3d in the bootstrap
procedure. }
\item{vectors.list}{A list of indices of eigenvectors used to build the projections in P.list. These indices are selected using the user specified model selection criterion as indicated in Steps 3a-3d in the bootstrap procedure. }
}
\description{
\loadmathjax
This function implements the first level of the parametric bootstrap procedure given by either Algorithm 1 or Algorithm 2 in Eck (2015) with respect to the mean-value parameterization. This is detailed in Steps 1 through 3d in the algorithm below.  This parametric bootstrap  generates resamples from the distribution evaluated at an envelope estimator of \mjeqn{\tau}{ascii} adjusting for model selection volatility.\cr
}
\details{
This function implements the first level of the parametric bootstrap
procedure given by either Algorithm 1 or Algorithm 2 in Eck (2015) with
respect to the mean-value parameterization. This is detailed in Steps 1
through 3d in the algorithm below.  This parametric bootstrap  generates
resamples from the distribution evaluated at an envelope estimator of \mjeqn{\tau}{ascii}
adjusting for model selection volatility.
The user specifies a model selection criterion which selects vectors that
construct envelope estimators using the reducing subspace approach. The user also
specifies which method is to be used in order to calculate envelope
estimators. When one is using a partial envelope, then this function
constructs envelope estimators of \mjeqn{\upsilon}{ascii} where we write \mjeqn{\tau}{ascii} = \mjeqn{(\gamma^T,\upsilon^T)^T}{ascii}
and \mjeqn{\upsilon}{ascii} corresponds to aster model parameters of interest.

In applications, candidate reducing subspaces are indices of eigenvectors of \mjeqn{ \widehat{\Sigma}_{\upsilon,\upsilon} }{ascii}
 where \mjeqn{ \widehat{\Sigma}_{\upsilon,\upsilon}}{ascii} is the part of \mjeqn{\hat{\Sigma}}{ascii}
corresponding to our parameters of interest. These indices are specified
by \code{vectors}. When all of the components of \mjeqn{\tau}{ascii} are components
of interest, then we write \mjeqn{\widehat{\Sigma}_{\upsilon,\upsilon} = \widehat{\Sigma} }{ascii}. When data
is generated via the parametric bootstrap, it is the indices (not the
                                                              original reducing subspaces) that are used to construct envelope estimators
constructed using the generated data. The algorithm using reducing subspaces
is as follows:

  \enumerate{
    \item Fit the aster model to the data and obtain
    \mjeqn{ \hat{\tau} = (\hat{\gamma}^T, \hat{\upsilon}^T) }{ascii} and \mjeqn{\hat{\Sigma}}{ascii}
   from the aster model fit.
    \item Compute the envelope estimator of \mjeqn{\upsilon}{ascii} in the original sample, given as
    \mjeqn{ \hat{\upsilon_{env}} = P_{\hat{G}}\hat{\upsilon}}{ascii} where \mjeqn{P_{\hat{G}}}{ascii} is computed using reducing subspaces
    and selected via a model selection criterion of choice.
    \item Perform a parametric bootstrap by generating resamples from the distribution of the
    aster submodel evaluated at \mjeqn{\hat{\tau}_{env} = (\hat{\gamma}^T,\hat{\upsilon_{env}}^T)^T}{ascii}. For iteration
    \mjeqn{b=1,...,B}{ascii} of the procedure:
      \enumerate{
        \item Compute \mjeqn{\hat{\tau}^{(b)}}{ascii} and \mjeqn{\hat{\Sigma}_{\upsilon,\upsilon}^{(b)}}{ascii} from the aster
        model fit to the resampled data.
        \item Build \mjeqn{P_{\hat{G}}^{(b)}}{ascii} using the indices of \mjeqn{\widehat{\Sigma}_{\upsilon,\upsilon}^{(b)}}{ascii} that
        are selected using the same model selection criterion as Step 2 to build \mjeqn{\hat{G}}{ascii}.
        \item Compute \mjeqn{\hat{\upsilon}_{env}^{(b)} = P_{\hat{\mathcal{E}}}^{(b)}\hat{\upsilon}^{(b)}}{ascii} and
        \mjeqn{\hat{\tau}_{env}^{(b)} = \left(\hat{\gamma}^{(b)^T},\hat{\upsilon}_{env}^{(b)^T}\right)^T}{ascii}.
        \item Store \mjeqn{\hat{\tau}_{env}^{(b)}}{ascii} and \mjeqn{g\left(\hat{\tau}_{env}^{(b)}\right)}{ascii}
        where \mjeqn{g}{ascii} maps \mjeqn{\tau}{ascii} to the parameterization of Darwinian fitness.
      }
    \item After \mjeqn{B}{ascii} steps, the bootstrap estimator of expected Darwinian fitness is the
    average of the envelope estimators stored in Step 3d. This completes the first part of the
    bootstrap procedure.
    \item We now proceed with the second level of bootstrapping at the \mjeqn{b^{th}}{ascii} stored envelope estimator
    \mjeqn{\hat{\tau}_{env}^{(b)}}{ascii}. For iteration \mjeqn{k=1,...,K}{ascii} of the procedure:
      \enumerate{
        \item Generate data from the distribution of the aster submodel evaluated at \mjeqn{\hat{\tau}_{env}^{(b)}}{ascii}.
        \item Perform Steps 3a through 3d with respect to the dataset obtained in Step 5a.
        \item Store \mjeqn{\hat{\tau}_{env}^{(b)^{(k)}}}{ascii} and \mjeqn{g\left(\hat{\tau}_{env}^{(b)^{(k)}}\right)}{ascii}.
      }
  }


The parametric bootstrap procedure which uses the 1d algorithm to construct
envelope estimators is analogous to the above algorithm. To use the 1d
algorithm, the user specifies \code{method = "1d"}. A parametric bootstrap
generating resamples from the distribution evaluated at the aster model
MLE is also conducted by this function.
}
\examples{
\dontrun{# see Web-based Supplementary Materials for ``Enveloping the aster model.''}
}
\references{
Cook, R.D. and Zhang, X. (2014). Foundations for Envelope Models and Methods. \emph{JASA}, In Press.\cr
\cr
Cook, R.D. and Zhang, X. (2015). Algorithms for Envelope Estimation. \emph{Journal of Computational and Graphical Statistics}, Published online. \doi{10.1080/10618600.2015.1029577}.\cr
\cr
Eck, D. J., Geyer, C. J., and Cook, R. D. (2016). Enveloping the aster model. \emph{in prep}. \cr
\cr
Eck, D.~J., Geyer, C.~J., and Cook, R.~D. (2016). Web-based Supplementary Materials for ``Enveloping the aster model.'' \emph{in prep}. \cr
\cr
Efron, B. (2014). Estimation and Accuracy After Model Selection. \emph{JASA}, \strong{109:507}, 991-1007.
}
