\name{generateddata}
\docType{data}
\alias{generateddata}
\title{A generated aster data set with 8 nodes}
\alias{redata}
\alias{data}
\alias{vars}
\description{
  Simulated data for an aster analysis. 
}

\usage{data(generateddata)}
\format{A dataframe with records for 400 organisms over 
  6 years. The description of objects included follows:

  \describe{

    \item{redata}{The data frame in long form. }
    \item{data}{The \code{asterdata} object. }
    \item{vars}{Character vector giving the names of the variables in 
      the graph. }
    \item{t}{Indicates survival for each of the 6 years. }
    \item{u}{Indicates flowering for each of the 6 years. }
    \item{v}{Counts flower heads for each of the 6 years. }
    \item{w}{Indicates seed production for each of the 6 years. }
    \item{y}{Counts seeds for each of the 6 years. }
    \item{z1}{The first trait thought to be associated with Darwinian 
      fitness. }
    \item{z2}{The second trait thought to be associated with 
      Darwinian fitness. }  
  }
}


%\source{}
%\references{}
\keyword{datasets}