\name{Mguttatus}
\docType{data}
\alias{Mguttatus}
\title{Mguttatus}
\alias{Mguttatus}
\alias{Mguttatus.redata}
\alias{quantities}
\description{
  A dataset with observed characteristics on 2313 Mimulus guttatus flowers.
}

\usage{data(Mguttatus)}
\format{A dataset with observed characteristics on 2313 Mimulus guttatus 
  flowers. The description of objects included follows:

  \describe{

    \item{Mguttatus.redata}{The data frame in long form. }
    \item{Mguttatus}{The \code{asterdata} object. }
    \item{quantities}{The quantities required to fit the aster model. The 
      components not described in the aster help section are described below.}
    \item{fit}{The indices of nodes in the \code{Mguttatus.redata} dataset 
      that correspond to Darwinian fitness. }
    \item{gen_bac}{The genetic background of the M. guttatus flower. }
    \item{site}{The field site of the M. guttatus flower. }
    \item{inversion}{The inversion orientation of the M. guttatus flower.}
    \item{ecotype}{The ecotype of the M. guttatus flower.}
  }
}


\references{
  Lowry, D.~B. and Willis, J.~H. (2010).
  A Widespread Chromosomal Inversion Polymorphism Contributes to a Major Life-History Transition, Local Adaptation, and Reproductive Isolation.
  \eqn{\emph{PLoS Biol}}, 8(9): e1000500. \doi{10.1371/journal.pbio.1000500}.
}

%\source{}
%\references{}
\keyword{datasets}
